Move the layer factory to the ol-map component so any app can benefit from using the simplified config
