+++ components/ol-map/LayerFactory.js
@@ -1,4 +1,4 @@
-/* jshint esnext: true */
+
 
 import ol from 'openlayers';
 import can from 'can';
